backend/requirements.txt
Django==3.0.7
djangorestframework==3.11.0
mysqlclient==1.4.6
djoser==2.0.3